requirements.txt
aniso8601==9.0.1
bcrypt==3.2.2
blinker==1.9.0
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==2.0.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
Flask-Cors==3.0.10
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
pytz==2024.2
setuptools==75.6.0
six==1.17.0
SQLAlchemy==1.4.23
typing_extensions==4.12.2
Werkzeug==2.0.1
WTForms==3.2.1
